Extract components to individual files
